Consulta_rnc/ApiRNC/requirements.txt
absl-py==2.1.0
aiohttp==3.9.3
aiosignal==1.3.1
altgraph==0.17.4
annotated-types==0.6.0
anyio==4.2.0
APScheduler==3.10.4
arrow==1.3.0
astunparse==1.6.3
attrs==23.2.0
Babel==2.14.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
binaryornot==0.4.4
bitarray==2.9.2
blackboxai==2.5
blinker==1.7.0
bs4==0.0.2
build==1.2.1
CacheControl==0.14.0
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cookiecutter==2.6.0
crashtest==0.4.1
cryptography==42.0.5
cycler==0.12.1
decorator==5.1.1
Deprecated==1.2.14
discord==2.3.2
discord.py==2.3.2
distlib==0.3.8
distro==1.9.0
dnspython==2.0.0
docutils==0.21.2
dulwich==0.21.7
et-xmlfile==1.1.0
fastapi==0.109.2
fastjsonschema==2.19.1
filelock==3.13.3
Flask==3.0.2
Flask-Cors==5.0.0
Flask-Limiter==3.7.0
Flask-SQLAlchemy==3.1.1
flatbuffers==24.3.25
flet==0.22.0
flet-core==0.22.0
flet-runtime==0.22.0
fonttools==4.48.1
fpdf==1.7.2
frozenlist==1.4.1
fsspec==2024.3.1
future==1.0.0
gast==0.6.0
geocoder==1.38.1
geographiclib==2.0
geopy==2.4.1
google-pasta==0.2.0
gradio_client==0.14.0
greenlet==3.0.3
grpcio==1.66.2
gunicorn==23.0.0
h11==0.14.0
h2==4.1.0
h5py==3.11.0
hpack==4.0.0
httpcore==0.17.3
httptools==0.6.1
httpx==0.24.1
huggingface-hub==0.22.2
Hypercorn==0.17.3
hyperframe==6.0.1
idna==3.6
importlib_resources==6.4.0
installer==0.7.0
intel-openmp==2021.4.0
ipwhois==1.2.0
itsdangerous==2.1.2
jaraco.classes==3.4.0
Jinja2==3.1.3
joblib==1.4.0
keras==3.3.3
keyring==24.3.1
Kivy==2.3.0
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.7.0
Kivy-Garden==0.1.5
kiwisolver==1.4.5
libclang==18.1.1
limits==3.12.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.2
mdurl==0.1.2
mkl==2021.4.0
ml-dtypes==0.4.0
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.8
mssql==1.0.1
multidict==6.0.5
namex==0.0.8
networkx==3.3
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
openai==1.13.3
openpyxl==3.1.2
opt_einsum==3.4.0
optree==0.11.0
ordered-set==4.1.0
packaging==23.2
pandas==2.2.1
passlib==1.7.4
pdfminer==20191125
pdfminer.six==20231228
pdfplumber==0.11.0
pdfreader==0.1.12
pefile==2023.2.7
pexpect==4.9.0
phonenumbers==8.13.30
pillow==10.2.0
pkginfo==1.10.0
platformdirs==4.2.1
plyer==2.1.0
poetry==1.8.2
poetry-core==1.9.0
poetry-plugin-export==1.7.1
priority==2.0.0
protobuf==4.25.5
psycopg2==2.9.9
ptyprocess==0.7.0
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.6.1
pydantic_core==2.16.2
pygame==2.5.2
Pygments==2.17.2
pyinstaller==6.4.0
pyinstaller-hooks-contrib==2024.1
PyJWT==2.8.0
pymssql==2.2.11
pyodbc==5.1.0
pyparsing==3.1.1
PyPDF2==3.0.1
pypdfium2==4.28.0
pyperclip==1.8.2
pypiwin32==223
pypng==0.20220715.0
pyproject_hooks==1.1.0
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PyQt6==6.7.0
PyQt6-Qt6==6.7.2
PyQt6-sip==13.6.0
PySide6==6.7.2
PySide6_Addons==6.7.2
PySide6_Essentials==6.7.2
python-dateutil==2.8.2
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
QDarkStyle==3.2.3
qrcode==7.4.2
QtPy==2.4.1
rapidfuzz==3.9.0
ratelim==0.1.6
regex==2024.9.11
repath==0.9.0
reportlab==4.1.0
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.0
scikit-learn==1.4.2
scipy==1.13.0
setuptools==69.1.1
shellingham==1.5.4
shiboken6==6.7.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.27
starlette==0.36.3
sympy==1.12
tabulate==0.9.0
tbb==2021.12.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
termcolor==2.4.0
text-unidecode==1.3
threadpoolctl==3.5.0
tkcalendar==1.6.1
tomlkit==0.12.4
torch==2.3.0
tqdm==4.66.2
trove-classifiers==2024.4.10
ttkthemes==3.2.2
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.0
uvicorn==0.27.1
virtualenv==20.26.1
waitress==3.0.0
watchdog==4.0.0
watchfiles==0.21.0
websocket-client==1.7.0
websockets==11.0.3
Werkzeug==3.0.1
wheel==0.44.0
window==0.0.3
wrapt==1.16.0
wsproto==1.2.0
wxPython==4.2.1
xgboost==2.1.1
yarl==1.9.4
